requirements.txt
appdirs==1.4.4
astroid==2.4.2
attrs==19.3.0
beautifulsoup4==4.9.1
black==19.10b0
click==7.1.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.1
opencv-python==4.4.0.42
pandas==1.1.0
pathspec==0.8.0
Pillow==7.2.0
pylint==2.5.3
pytesseract==0.3.5
python-dateutil==2.8.1
pytz==2020.1
regex==2020.7.14
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
wrapt==1.12.1
